+++ tests/testthat/test-do_LigandReceptorPlot.R
@@ -1,58 +1,332 @@
 if(base::isFALSE(dep_check[["do_LigandReceptorPlot"]])){
   testthat::test_that("do_LigandReceptorPlot: CRAN essentials", {
-    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output)
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE)
     testthat::expect_type(p, "list")
   })
 
   testthat::test_that("do_LigandReceptorPlot: PASS - from output", {
     testthat::skip_on_cran()
     
-    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, dot_border = TRUE, use_viridis = TRUE, viridis.direction = 1)
+    suppressMessages({testthat::expect_message({p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = TRUE)})})
     testthat::expect_type(p, "list")
     
-    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, dot_border = TRUE, use_viridis = TRUE, viridis.direction = -1)
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "A", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
     testthat::expect_type(p, "list")
     
-    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, dot_border = TRUE, use_viridis = FALSE, sequential.direction = -1)
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "A", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
     testthat::expect_type(p, "list")
     
-    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, dot_border = TRUE, use_viridis = FALSE, sequential.direction = 1)
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "A", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
     testthat::expect_type(p, "list")
     
-    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, dot_border = FALSE, use_viridis = TRUE, viridis.direction = 1)
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "A", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
     testthat::expect_type(p, "list")
     
-    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, dot_border = FALSE, use_viridis = TRUE, viridis.direction = -1)
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "A", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
     testthat::expect_type(p, "list")
     
-    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, dot_border = FALSE, use_viridis = FALSE, sequential.direction = -1)
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "A", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
     testthat::expect_type(p, "list")
     
-    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, dot_border = FALSE, use_viridis = FALSE, sequential.direction = 1)
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "A", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
     testthat::expect_type(p, "list")
     
-    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, plot.grid = TRUE)
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "A", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
     testthat::expect_type(p, "list")
     
-    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, plot.grid = FALSE)
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "A", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
     testthat::expect_type(p, "list")
     
-    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, plot.grid = TRUE, dot_border = FALSE)
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "A", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "A", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "A", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "A", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "A", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "A", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "A", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    
+    
+    
+    
+    
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "B", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "B", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "B", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "B", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "B", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "B", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "B", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "B", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "B", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "B", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "B", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "B", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "B", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "B", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "B", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "B", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "C", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "C", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "C", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "C", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "C", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "C", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "C", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "C", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "C", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "C", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "C", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "C", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "C", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "C", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "C", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "C", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    
+    
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "D", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "D", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "D", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "D", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "D", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "D", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "D", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "D", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "D", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "D", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "D", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "D", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "D", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "D", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "D", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "D", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    
+    
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "E", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "E", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "E", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "E", invert_specificity = TRUE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "E", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "E", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "E", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "E", invert_specificity = TRUE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "E", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "E", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "E", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "E", invert_specificity = FALSE, invert_magnitude = TRUE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "E", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "E", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "ascending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "E", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "ascending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE, sort.by = "E", invert_specificity = FALSE, invert_magnitude = FALSE, sorting.type.specificity = "descending", sorting.type.magnitude = "descending")
+    testthat::expect_type(p, "list")
+    
+    
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, dot_border = TRUE, use_viridis = TRUE, viridis.direction = 1, verbose = FALSE)
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, dot_border = TRUE, use_viridis = TRUE, viridis.direction = -1, verbose = FALSE)
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, dot_border = TRUE, use_viridis = FALSE, sequential.direction = -1, verbose = FALSE)
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, dot_border = TRUE, use_viridis = FALSE, sequential.direction = 1, verbose = FALSE)
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, dot_border = FALSE, use_viridis = TRUE, viridis.direction = 1, verbose = FALSE)
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, dot_border = FALSE, use_viridis = TRUE, viridis.direction = -1, verbose = FALSE)
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, dot_border = FALSE, use_viridis = FALSE, sequential.direction = -1, verbose = FALSE)
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, dot_border = FALSE, use_viridis = FALSE, sequential.direction = 1, verbose = FALSE)
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, plot.grid = TRUE, verbose = FALSE)
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, plot.grid = FALSE, verbose = FALSE)
+    testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, plot.grid = TRUE, dot_border = FALSE, verbose = FALSE)
     testthat::expect_type(p, "list")
 
-    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output)
+    p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, verbose = FALSE)
     testthat::expect_type(p, "list")
 
     p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output, 
                                        keep_source = c("NK", "B"),
-                                       keep_target = "CD8 T")
+                                       keep_target = "CD8 T", verbose = FALSE)
     testthat::expect_type(p, "list")
   })
 
   testthat::test_that("do_LigandReceptorPlot: PASS - from output different n", {
     testthat::skip_on_cran()
     p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output,
-                                       top_interactions = 50)
+                                       top_interactions = 50, verbose = FALSE)
     testthat::expect_type(p, "list")
   })
 
@@ -60,26 +334,26 @@ if(base::isFALSE(dep_check[["do_LigandReceptorPlot"]])){
   testthat::test_that("do_LigandReceptorPlot: PASS - split.by", {
     testthat::skip_on_cran()
     p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output,
-                                       split.by = "ligand.complex")
+                                       split.by = "ligand.complex", verbose = FALSE)
     testthat::expect_type(p, "list")
 
     p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output,
-                                       split.by = "receptor.complex")
+                                       split.by = "receptor.complex", verbose = FALSE)
     testthat::expect_type(p, "list")
   })
 
   testthat::test_that("do_LigandReceptorPlot: PASS - from output, angle ", {
     testthat::skip_on_cran()
     p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output,
-                                       axis.text.x.angle = 0)
+                                       axis.text.x.angle = 0, verbose = FALSE)
     testthat::expect_type(p, "list")
 
     p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output,
-                                       axis.text.x.angle = 45)
+                                       axis.text.x.angle = 45, verbose = FALSE)
     testthat::expect_type(p, "list")
 
     p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output,
-                                       axis.text.x.angle = 90)
+                                       axis.text.x.angle = 90, verbose = FALSE)
     testthat::expect_type(p, "list")
   })
 
@@ -88,12 +362,12 @@ if(base::isFALSE(dep_check[["do_LigandReceptorPlot"]])){
     testthat::skip_on_cran()
     p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output,
                                        top_interactions = 50,
-                                       legend.position = "bottom")
+                                       legend.position = "bottom", verbose = FALSE)
     testthat::expect_type(p, "list")
 
     p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output,
                                        top_interactions = 50,
-                                       legend.position = "right")
+                                       legend.position = "right", verbose = FALSE)
     testthat::expect_type(p, "list")
   })
 
@@ -104,37 +378,37 @@ if(base::isFALSE(dep_check[["do_LigandReceptorPlot"]])){
     testthat::skip_on_cran()
     p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output,
                                        top_interactions = 50,
-                                       sort_interactions_alphabetically =  TRUE)
+                                       sort_interactions_alphabetically =  TRUE, verbose = FALSE)
     testthat::expect_type(p, "list")
 
     p <- SCpubr::do_LigandReceptorPlot(liana_output = liana_output,
                                        top_interactions = 50,
-                                       sort_interactions_alphabetically =  FALSE)
+                                       sort_interactions_alphabetically =  FALSE, verbose = FALSE)
     testthat::expect_type(p, "list")
   })
 
   testthat::test_that("do_LigandReceptorPlot: FAIL - wrong parameters", {
     testthat::skip_on_cran()
     testthat::expect_error({SCpubr::do_LigandReceptorPlot(liana_output = liana_output,
-                                                          font.type = "wrong")})
+                                                          font.type = "wrong", verbose = FALSE)})
 
     testthat::expect_error({SCpubr::do_LigandReceptorPlot(liana_output = liana_output,
-                                                          legend.type = "wrong")})
+                                                          legend.type = "wrong", verbose = FALSE)})
 
     testthat::expect_error({SCpubr::do_LigandReceptorPlot(liana_output = liana_output,
-                                                          axis.text.x.angle = 10)})
+                                                          axis.text.x.angle = 10, verbose = FALSE)})
 
     testthat::expect_error({SCpubr::do_LigandReceptorPlot(liana_output = liana_output,
-                                                          font.type = "wrong")})
+                                                          font.type = "wrong", verbose = FALSE)})
 
     testthat::expect_error({SCpubr::do_LigandReceptorPlot(liana_output = liana_output,
-                                                          legend.position = "wrong")})
+                                                          legend.position = "wrong", verbose = FALSE)})
 
     testthat::expect_error({SCpubr::do_LigandReceptorPlot(liana_output = liana_output,
-                                                          grid.type = "wrong")})
+                                                          grid.type = "wrong", verbose = FALSE)})
 
     testthat::expect_error({SCpubr::do_LigandReceptorPlot(liana_output = liana_output,
-                                                          split.by = "wrong")})
+                                                          split.by = "wrong", verbose = FALSE)})
 
   })
 }